Supported x86_64.
git-svn-id: https://gl-xemmai.googlecode.com/svn/trunk@3 e18edb27-47f2-8e31-856d-7dbf3ad89f5d

+++ gl/xemmai/program.h
@@ -64,7 +64,7 @@ public:
 		GLenum type;
 		std::vector<GLchar> name(n);
 		glGetActiveAttrib(f_id(), a_index, n, NULL, &size, &type, &name[0]);
-		return f_tuple(t_transfer(size), t_transfer(static_cast<int>(type)), f_global()->f_as(f_convert(&name[0])));
+		return f_tuple(t_transfer(size), t_transfer(type), f_global()->f_as(f_convert(&name[0])));
 	}
 	t_transfer f_get_active_uniform(GLuint a_index) const
 	{
@@ -73,7 +73,7 @@ public:
 		GLenum type;
 		std::vector<GLchar> name(n);
 		glGetActiveUniform(f_id(), a_index, n, NULL, &size, &type, &name[0]);
-		return f_tuple(t_transfer(size), t_transfer(static_cast<int>(type)), f_global()->f_as(f_convert(&name[0])));
+		return f_tuple(t_transfer(size), t_transfer(type), f_global()->f_as(f_convert(&name[0])));
 	}
 	t_transfer f_get_attached_shaders() const;
 	GLint f_get_attrib_location(const std::wstring& a_name) const
